=act split Props java api implementation in separate files.
